make dynamicStyles tests use hooks
